Put auto towers in a separate file
